Got stuff straightened out.
At first, I thought I did a commit on the wrong file.  What it actually was, I forgot to test the script on a first run before uploading it.  The were a some files that needed to be written for the first run. They were missing and that cause  errors.  Ran shellcheck on it.  The majority of shellcheck's issues:  1. all variables needed to be double quoted when they didn't all need to be  double quoted.  2. all read commands needing the -r option when they didn't all need the -r option.  I made those changes to cut down on the amount of output from shellcheck.   3.Another big thing was shellcheck  report there should be  || exit  added to the cd command. If the cd command failed.  That would exit the user out of the script if the user entered an empty string or  invalid directory.  This script does the majority of the error handling.   If the user enters a blank string or invalid directory  the script displays those error messages to the user.  There is a check done to make sure the directory is valid before the cd command is applied to the users entry.  If the directory doesn't exist, the cd command isn't executed. I added play notifications with the filename of what was played.  The path was stripped out so only the filename shows in the notifications. Next on the list is error handling for the play notifications.  Right now if an attempt is made to play an invalid file the notification "Playing.." comes up.  It needs to notify the user that it was and invalid file.