Major Layout changes to remove interface redundancy. + more features from v1.99 to v2.00
Major Layout changes to remove interface redundancy. + more features v.2.00
Now home, playlist and played list screens are aware if the clipboard has a valid file  or a valid url link in it and act accordingly.  Yt-dlp now tries to extract a fliename from the link in the clipboard.  If it is successful, the videos name will be displayed.  Yt-dlp was hanging sometimes on invalid urls while trying to extract a vdeo's name.  So, a setting was added for a yt-dlp timeout.  The default setting is 5 seconds.  That can be changed from <YT-DLP>  Now every time there is a valid file or a valid url in the clipboard... <Play Clipboard> will blink to let the user know there is playable content in the clipboard.

Next update will also make the  <YT-DLP.>   screen aware if the content in the clipboard is a file or url.  Then onto playlist enhancements,  

Help hasn't been updated yet for this ver.