v2.04 release.  Play Clipboard now on all content screens
Squashed a bug in Play Clipboard.  That feature would fail when url had a video filename that had a dot before the .ext

New feature added in  >>> SETTINGS >>> LAST SEARCH 

If there are search results on the the home screen on exit, the below always applies to your next run.

1 - Always Save
2 - Never Save
3 - Ask on Exit

The next version release v2.05 will only be the Help update.