v2.068 Bug Squashed in Play Clipboard and Last Download.
Fixed auto play failing with urls in Play Clipboard. This issue also caused Last Download auto play to fail. 

Below is what would cause the failure:

Only with a URL. If you download a video file and you already have a filename with the same name but a
different video extension in the current directory, it would cause Play Clipboard auto play to fail.
Last Download would also fail to play in this situation.