v2.8.10 Added: Command line option -u
This option can be used with your own ready to download url list that can be any name and location.
There is no need to use the Videoinfox interface with this option. Videos will be downloaded to the
directory of the list location.  Example: videoinfox -u mylist  Only enter the list filename.
The list must be recursively reachable from the current directory.
The first line in the list must begin with http to be processed.

videoinfox -h  will display the above instructions.