v3.5.08 Fix in VIEW PLAYLIST >>> CHANGE PLAYLIST - Change Playlist when the DEFAULT Playlist was deleted.
When the Default Playlist was deleted it generated a script error from the cp command.

The Default Playlist can't be deleted. It can only be cleared. If the Default Playlist is deleted, it is
temporarily removed (not deleted) from CHANGE PLAYLIST. The Default Playlist will show again in Change Playlist
after another playlist is loaded. This issue in Change Playlist will be dealt with in a future update.