v4.1.12 Change: Blinking dash in Ytd-lp menu options -Download Default -Download Choose -Self Destruct changed to ~
Change: Clipboard was saved when entering Yt-dlp >>> Download List and restored on quit even if the clipboard
was cleared while in Yt-dlp >>> Download List -Default List. Now the clipboard isn't restored if cleared.

Fix: The default playlist could be deleted.
Message: The default playlist list can't be deleted. It can only be cleared . . .