v4.7.00 Add: --save-position-on-quit to mpv options.
When Quit(q) is used to exit Mpv, the current play position will be saved for every video played with Mpv.
When a video is played that Mpv has a position marker for, the video will start playing from that marker.
Mpv handles the markers for the videos. Located in ~/.config/mpv/watch_later

Add: -Mpv Playlist to VIEW PLAYLIST >>> CHANGE AND QUEUE
The above option will load a playlist into Mpv. This will bypass the -Resume Autoplay option and the
Intermission Screen in Videoinfox. So when -Mpv Playlist is used, it will keep track of playlist position
instead of Videoinfox's Resume Autoplay option.

Added to Home Help * If Mpv is used, it will keep track of last play position on all videos played on Mpv quit(q)

Added -Mpv Playlist info to View Playlist >>> Help