v4.7.01 Add: Below added to Playlist >>> Help
Note: This applies to only to mpv . . .
If a video is quit(q) while playing in mpv, it will create a marker of the last play position.

If using Mpv Playlist and back track(<) or forward track(>) are selected, the mpv marker will be
erased for the video that was playing (if it had a marker) when < or > were selected. To avoid
any mpv markers from being erased use -Resume Autoplay instead of -Mpv Playlist to watch playlists.