v5.4.26 Add: Below info added to Home Help in Default Directories section.
Set default directories first then -Index Default Directories located in settings.
Every time a default directory is changed or removed in settings, or if any content has been added or removed
from any of the default directories, Index Default Directories will need to be run again to update the changes.
The above also applies if any videos are renamed in any of the default directories. A date and time stamp is
displayed in settings of the last time Index Default Directories was run.