v5.7.00 New: Now when adding a URL Playlist there is an option to customize the Yt-dlp options for each playlist.
New: -Set in -Browse. Each domain has it own link format. This is the info in the link after the .com but
before the video id. This is needed for browse to download single videos from the playlist. Other sites 
besides Youtube can now be be used with Download Playlist and Browse.

 Add: -Delete in -Download Playlist.

Fix: Had to fix some wonky code with restore.

Edit: Changed the prompt messages for -Remove and -Restore to provide more info. on how it works.